frontend/requirements.txt
streamlit
requests
python-docx
pdfplumber
